refactor project structure/ remove unused classes
