+++ app/src/main/java/com/derek_s/hubble_gallery/api/GetAlbum.java
@@ -3,7 +3,7 @@ package com.derek_s.hubble_gallery.api;
 import android.os.AsyncTask;
 
 import com.derek_s.hubble_gallery._shared.model.TileObject;
-import com.derek_s.hubble_gallery.nav_drawer.fragments.FragMain;
+import com.derek_s.hubble_gallery.ui.fragments.FragMain;
 
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
