+++ app/src/main/java/com/derek_s/hubble_gallery/ui/widgets/CircularProgressDrawable.java
@@ -1,4 +1,4 @@
-package com.derek_s.hubble_gallery.nav_drawer.ui.widgets;
+package com.derek_s.hubble_gallery.ui.widgets;
 
 import android.animation.Animator;
 import android.animation.ArgbEvaluator;
