--- app/src/main/java/com/derek_s/hubble_gallery/utils/NavDataUtils.java
@@ -1,22 +1,0 @@
-package com.derek_s.hubble_gallery.utils;
-
-import com.derek_s.hubble_gallery.nav_drawer.model.SectionChildObject;
-import com.derek_s.hubble_gallery.nav_drawer.model.SectionObject;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-
-public class NavDataUtils {
-
-    public static ArrayList<SectionObject> addAllGroups() {
-        ArrayList<SectionObject> result = new ArrayList<>();
-
-        return result;
-    }
-
-    public static HashMap<String, ArrayList<SectionChildObject>> getSectionChildren() {
-        HashMap<String, ArrayList<SectionChildObject>> result = new HashMap<>();
-
-        return result;
-    }
-}
