+++ app/src/main/res-details/menu/menu_act_details.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".nav_drawer.ui.activities.ActDetails">
+    tools:context=".ui.activities.ActDetails">
 
     <item
         android:id="@+id/action_expand"
